+++ R/create_edges_spectra.R
@@ -11,6 +11,7 @@
 #' @param output Output file.
 #' @param name_source Name of the source features column
 #' @param name_target Name of the target features column
+#' @param method Similarity method
 #' @param threshold Minimal similarity to report
 #' @param ppm Relative ppm tolerance to be used
 #' @param dalton Absolute Dalton tolerance to be used
@@ -35,7 +36,8 @@ create_edges_spectra <- function(input = get_params(step = "create_edges_spectra
                                  output = get_params(step = "create_edges_spectra")$files$networks$spectral$edges$raw,
                                  name_source = get_params(step = "create_edges_spectra")$names$source,
                                  name_target = get_params(step = "create_edges_spectra")$names$target,
-                                 threshold = get_params(step = "create_edges_spectra")$annotations$thresholds$ms2$similarity$edges,
+                                 method = get_params(step = "create_edges_spectra")$similarities$methods$edges,
+                                 threshold = get_params(step = "create_edges_spectra")$similarities$thresholds$edges,
                                  ppm = get_params(step = "create_edges_spectra")$ms$tolerances$mass$ppm$ms2,
                                  dalton = get_params(step = "create_edges_spectra")$ms$tolerances$mass$dalton$ms2,
                                  qutoff = get_params(step = "create_edges_spectra")$ms$thresholds$ms2$intensity) {
@@ -56,10 +58,13 @@ create_edges_spectra <- function(input = get_params(step = "create_edges_spectra
     log_debug("Take yourself a break, you deserve it.")
     nspecz <- length(spectra)
     fragz <- spectra@backend@peaksData
+    precz <- spectra@backend@spectraData$precursorMz
 
     edges <- create_edges(
       frags = fragz,
       nspecs = nspecz,
+      precs = precz,
+      method = method,
       ms2_tolerance = dalton,
       ppm_tolerance = ppm,
       threshold = threshold